refactor: Location of files changed according to project policy #36
