+++ src/app/components/pricing-plan/pricing-plan.component.scss
