+++ src/app/components/user-management/login/login.component.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @angular-eslint/no-empty-lifecycle-method */
 import { Component, OnInit } from '@angular/core';
-import { LoginUser } from './models/userLogin.models';
-import { LoginService } from './login.service';
+import { LoginUser } from '../../../models/userLogin.models';
+import { LoginService } from '../../../service/login.service';
 
 
 @Component({
