+++ src/app/components/user-management/register/register.component.scss
