+++ src/app/modules/login.module.ts
@@ -2,7 +2,7 @@ import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule } from '@angular/forms';
-import { LoginComponent } from './login.component';
+import { LoginComponent } from '../components/user-management/login/login.component';
 
 @NgModule({
   declarations: [LoginComponent],
