+++ src/app/modules/logout.module.ts
@@ -2,8 +2,8 @@ import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule } from '@angular/forms';
-import { LogoutComponent } from './logout.component';
-import { Logout } from './logout.service';
+import { LogoutComponent } from '../components/user-management/logout/logout.component';
+import { Logout } from '../service/logout.service';
 
 @NgModule({
   declarations: [LogoutComponent],
