+++ src/app/modules/pricing-plan.module.ts
@@ -2,8 +2,8 @@ import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule } from '@angular/forms';
-import { PricingPlanComponent } from './pricing-plan.component';
-import { MemberType } from './models/pricing.models';
+import { PricingPlanComponent } from '../components/pricing-plan/pricing-plan.component';
+import { MemberType } from '../models/pricing.models';
 
 @NgModule({
   declarations: [PricingPlanComponent],
