--- src/app/pricing-plan/pricing-plan.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PricingPlanService } from './pricing-plan.service';
-
-describe('PricingPlanService', () => {
-  let service: PricingPlanService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PricingPlanService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
