--- src/app/tab2/auth-choice/auth-choice.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { AuthChoiceComponent } from './auth-choice.component';
-
-describe('AuthChoiceComponent', () => {
-  let component: AuthChoiceComponent;
-  let fixture: ComponentFixture<AuthChoiceComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ AuthChoiceComponent ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(AuthChoiceComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
