--- src/app/usuarios/logout/logout.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { LogoutService } from './logout.service';
-
-describe('LogoutService', () => {
-  let service: LogoutService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(LogoutService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
