--- src/app/usuarios/logout/models/userLogout.models.ts
@@ -1,4 +1,0 @@
-export interface LogoutUser {
-  email: string;
-  password: string;
-}
