--- src/app/usuarios/register/register.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { RegisterService } from './register.service';
-
-describe('RegisterService', () => {
-  let service: RegisterService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(RegisterService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
